Get basic auth flow working.
Right now, we're pulling from the database on every call. This is probably fine
because for now it's going on loopback interface over my laptop. This should be
fast enough for now in the happy path. In the future, we might want to cache the
key in memory for a speedup. Ideally, it should only have to be loaded once on startup.